refactoring Add view all or by id Handle not found records exceptions

+++ src/main/java/gamz/projects/pharmacyfair/model/request/FeedbackProcessRequest.java
@@ -14,5 +14,5 @@ import lombok.NoArgsConstructor;
 public class FeedbackProcessRequest {
     @Valid
     @NotEmpty(message = "Вы должны указать номер заявки")
-    private Integer feedbackId;
+    private Long feedbackId;
 }
