+++ src/main/java/gamz/projects/pharmacyfair/repository/FeedbackRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import gamz.projects.pharmacyfair.model.entity.Feedback;
 
-public interface FeedbackRepository extends CrudRepository<Feedback,Integer> {
+public interface FeedbackRepository extends CrudRepository<Feedback,Long> {
     
 }
